Still a bit borked... no stream with pi cam????

+++ photon-core/src/main/java/org/photonvision/vision/pipeline/CVPipelineSettings.java
@@ -42,7 +42,7 @@ public class CVPipelineSettings implements Cloneable {
     public String pipelineNickname = "New Pipeline";
     public boolean cameraAutoExposure = false;
     // manual exposure only used if cameraAutoExposure is false
-    public double cameraExposure = 20;
+    public int cameraExposure = 20;
     public int cameraBrightness = 50;
     // Currently only used by a few cameras (notably the zero-copy Pi Camera driver) with the Gain
     // quirk
